# server/api/v1/web/web_homepage.go
package web

import (
	"github.com/flipped-aurora/gin-vue-admin/server/global"
	"github.com/flipped-aurora/gin-vue-admin/server/model/common/request"
	"github.com/flipped-aurora/gin-vue-admin/server/model/common/response"
	"github.com/gin-gonic/gin"
	"go.uber.org/zap"
)

type HomePageApi struct {
}

// GetOrganizationList
// @Tags      HomePage
// @Summary   分页获取社团列表
// @Security  ApiKeyAuth
// @accept    application/json
// @Produce   application/json
// @Param     data query     request.PageInfo                                        true  "页码, 每页大小"
// @Success   200   {object}  response.Response{data=response.PageResult,msg=string}  "分页获取社团列表,返回包括列表,总数,页码,每页数量"
// @Router    /homepage/organizationList [get]
func (homePageApi *HomePageApi) GetOrganizationList(c *gin.Context) {

	var pageInfo request.PageInfo
	err := c.ShouldBindQuery(&pageInfo)
	if err != nil {
		response.FailWithMessage(err.Error(), c)
		return
	}
	if list, total, err := userService.GetOrganizationList(pageInfo); err != nil {
		global.GVA_LOG.Error("获取失败!", zap.Error(err))
		response.FailWithMessage("获取失败", c)
	} else {
		response.OkWithDetailed(response.PageResult{
			List:     list,
			Total:    total,
			Page:     pageInfo.Page,
			PageSize: pageInfo.PageSize,
		}, "获取成功", c)
	}
}