# server/docs/swagger.yaml
definitions:
...
  request.IdsReq:
    properties:
      ids:
        items:
          type: integer
        type: array
    type: object
...
  response.PageResult:
    properties:
      list: {}
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  web.Activity:
    properties:
      beginTime:
        type: string
      content:
        type: string
      cover:
        type: string
      createdAt:
        description: 创建时间
        type: string
      createdBy:
        type: integer
      deletedBy:
        type: integer
      endTime:
        type: string
      id:
        description: 主键ID
        type: integer
      status:
        type: string
      title:
        type: string
      type:
        type: integer
      updatedAt:
        description: 更新时间
        type: string
      updatedBy:
        type: integer
      userId:
        type: integer
    type: object
...
  web.Article:
    properties:
      categoryId:
        type: integer
      content:
        type: string
      cover:
        type: string
      createdAt:
        description: 创建时间
        type: string
      id:
        description: 主键ID
        type: integer
      isTop:
        type: boolean
      status:
        type: integer
      title:
        type: string
      uid:
        type: integer
      updatedAt:
        description: 更新时间
        type: string
      viewNum:
        type: integer
    type: object
...
info:
  contact: {}
  description: This is a sample Server pets
  title: Swagger Example API
  version: 0.0.1
paths:
  /activity/createActivity:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建Activity
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/web.Activity'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 创建Activity
      tags:
      - Activity
  /activity/deleteActivity:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 删除Activity
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/web.Activity'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 删除Activity
      tags:
      - Activity
  /activity/deleteActivityByIds:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 批量删除Activity
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.IdsReq'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"批量删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 批量删除Activity
      tags:
      - Activity
  /activity/findActivity:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: beginTime
        type: string
      - in: query
        name: content
        type: string
      - in: query
        name: cover
        type: string
      - description: 创建时间
        in: query
        name: createdAt
        type: string
      - in: query
        name: createdBy
        type: integer
      - in: query
        name: deletedBy
        type: integer
      - in: query
        name: endTime
        type: string
      - description: 主键ID
        in: query
        name: id
        type: integer
      - in: query
        name: status
        type: string
      - in: query
        name: title
        type: string
      - in: query
        name: type
        type: integer
      - description: 更新时间
        in: query
        name: updatedAt
        type: string
      - in: query
        name: updatedBy
        type: integer
      - in: query
        name: userId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"查询成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 用id查询Activity
      tags:
      - Activity
  /activity/getActivityList:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: beginTime
        type: string
      - in: query
        name: content
        type: string
      - in: query
        name: cover
        type: string
      - description: 创建时间
        in: query
        name: createdAt
        type: string
      - in: query
        name: createdBy
        type: integer
      - in: query
        name: deletedBy
        type: integer
      - in: query
        name: endCreatedAt
        type: string
      - in: query
        name: endTime
        type: string
      - description: 主键ID
        in: query
        name: id
        type: integer
      - description: 关键字
        in: query
        name: keyword
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - in: query
        name: startCreatedAt
        type: string
      - in: query
        name: status
        type: string
      - in: query
        name: title
        type: string
      - in: query
        name: type
        type: integer
      - description: 更新时间
        in: query
        name: updatedAt
        type: string
      - in: query
        name: updatedBy
        type: integer
      - in: query
        name: userId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 分页获取Activity列表
      tags:
      - Activity
  /activity/updateActivity:
    put:
      consumes:
      - application/json
      parameters:
      - description: 更新Activity
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/web.Activity'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"更新成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 更新Activity
      tags:
      - Activity
...
  /article/createArticle:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建Article
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/web.Article'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 创建Article
      tags:
      - Article
  /article/deleteArticle:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 删除Article
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/web.Article'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 删除Article
      tags:
      - Article
  /article/deleteArticleByIds:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 批量删除Article
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.IdsReq'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"批量删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 批量删除Article
      tags:
      - Article
  /article/findArticle:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: categoryId
        type: integer
      - in: query
        name: content
        type: string
      - in: query
        name: cover
        type: string
      - description: 创建时间
        in: query
        name: createdAt
        type: string
      - description: 主键ID
        in: query
        name: id
        type: integer
      - in: query
        name: isTop
        type: boolean
      - in: query
        name: status
        type: integer
      - in: query
        name: title
        type: string
      - in: query
        name: uid
        type: integer
      - description: 更新时间
        in: query
        name: updatedAt
        type: string
      - in: query
        name: viewNum
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"查询成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 用id查询Article
      tags:
      - Article
  /article/getArticleList:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: categoryId
        type: integer
      - in: query
        name: content
        type: string
      - in: query
        name: cover
        type: string
      - description: 创建时间
        in: query
        name: createdAt
        type: string
      - in: query
        name: endCreatedAt
        type: string
      - description: 主键ID
        in: query
        name: id
        type: integer
      - in: query
        name: isTop
        type: boolean
      - description: 关键字
        in: query
        name: keyword
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - in: query
        name: startCreatedAt
        type: string
      - in: query
        name: status
        type: integer
      - in: query
        name: title
        type: string
      - in: query
        name: uid
        type: integer
      - description: 更新时间
        in: query
        name: updatedAt
        type: string
      - in: query
        name: viewNum
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 分页获取Article列表
      tags:
      - Article
  /article/updateArticle:
    put:
      consumes:
      - application/json
      parameters:
      - description: 更新Article
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/web.Article'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"更新成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 更新Article
      tags:
      - Article
...
  /homepage/organizationList:
    get:
      consumes:
      - application/json
      parameters:
      - description: 关键字
        in: query
        name: keyword
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 分页获取社团列表,返回包括列表,总数,页码,每页数量
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PageResult'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 分页获取社团列表
      tags:
      - HomePage
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"